requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-crispy-forms==1.7.2
idna==2.8
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
pipenv==2018.11.26
psutil==5.6.2
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
